+++ cmd/firesol/cli/common-flag.go
@@ -19,7 +19,7 @@ func init() {
 		// Network config
 		cmd.Flags().String("common-network-id", NetworkID, "[COMMON] Solana network identifier known to us for pre-configured elements Used by: miner-node, reader-node")
 		cmd.Flags().String("common-sf-network-id", SFNetworkID, "[COMMON] StreamingFast network ID, used for some billing functions by dgraphql")
-		cmd.Flags().Uint64("common-protocol-first-streamable-block", 0, "[COMMON] first chain streamable block. Not genesis")
+		cmd.Flags().Uint64("common-first-streamable-block", 0, "[COMMON] first chain streamable block. Not genesis")
 
 		cmd.Flags().Bool("common-blocks-cache-enabled", false, FlagDescription(`
 				[COMMON] Use a disk cache to store the blocks data to disk and instead of keeping it in RAM. By enabling this, block's Protobuf content, in bytes,